requirements.txt
asgiref==3.4.1
boto3==1.19.12
botocore==1.22.12
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.12
django-allauth==0.50.0
django-appconf==1.0.5
django-compressor==2.4.1
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-import-export==2.6.1
django-registration==3.2
django-rest-auth==0.9.5
django-sass-processor==1.1
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
djangorestframework-jwt==1.11.0
drf-access-policy==1.1.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jmespath==0.10.0
libsass==0.21.0
MarkupPy==1.14
MarkupSafe==2.0.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.2
Pillow==8.3.2
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
rcssmin==1.0.6
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
tablib==3.1.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==6.1.0
xlrd==2.0.1
xlwt==1.3.0
